Refactor challenge scripts
